desloppify: split rimap-content::html into style_parse/hidden/mismatch/extract/sanitize
