+++ crates/rimap-content/src/html/mod.rs
@@ -9,13 +9,22 @@
 //! The single public (crate-visible) entrypoint is [`process`].
 
 use std::collections::HashSet;
-use std::sync::LazyLock;
 
-use ammonia::Builder;
-use scraper::{Html, Selector};
+use scraper::Html;
 
 use crate::error::ContentError;
 use crate::output::SecurityWarning;
+
+mod extract;
+mod hidden;
+mod mismatch;
+mod sanitize;
+mod style_parse;
+
+use crate::html::extract::{collect_anchor_hrefs, extract_text};
+use crate::html::hidden::detect_hidden;
+use crate::html::mismatch::detect_mismatches;
+use crate::html::sanitize::sanitize_body;
 
 /// Result of processing a single HTML body part.
 #[derive(Debug, Clone)]
@@ -50,7 +59,7 @@ pub(crate) const MAX_MISMATCH_HITS: usize = 32;
 pub(crate) const HTML_BODY_LIMIT_KIND: &str = "html_body";
 
 /// Stable identifier for an element we've decided is hidden. Used by
-/// `extract_text` (Task 9) to skip hidden subtrees.
+/// `extract_text` to skip hidden subtrees.
 ///
 /// `scraper` does not give us a stable `ElementRef` across re-parses,
 /// so we identify hidden elements by their position in a pre-order
@@ -62,7 +71,7 @@ pub(crate) type ElementIndex = usize;
 ///
 /// Each variant corresponds to a distinct CSS pattern that effectively
 /// removes content from rendering, used by hidden-content detection in
-/// [`detect_hidden`].
+/// [`crate::html::hidden::detect_hidden`].
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub(crate) enum HiddenMethod {
     /// `display: none` — element omitted from layout entirely.
@@ -93,506 +102,6 @@ impl HiddenMethod {
     }
 }
 
-/// Compile a const CSS selector string. The `expect` is unreachable for
-/// const selector inputs and is exercised at first use of each
-/// `LazyLock` so the lint expectation is fulfilled.
-#[expect(
-    clippy::expect_used,
-    reason = "const CSS selector strings cannot fail at runtime"
-)]
-fn compile_selector(src: &'static str) -> Selector {
-    Selector::parse(src).expect("rimap-content: invalid const CSS selector")
-}
-
-/// Selector matching anchor elements with an `href` attribute.
-static SEL_ANCHOR: LazyLock<Selector> = LazyLock::new(|| compile_selector("a[href]"));
-/// Selector matching `<img>` elements.
-static SEL_IMG: LazyLock<Selector> = LazyLock::new(|| compile_selector("img"));
-/// Selector matching `<script>` elements.
-static SEL_SCRIPT: LazyLock<Selector> = LazyLock::new(|| compile_selector("script"));
-/// Selector matching `<style>` elements.
-static SEL_STYLE: LazyLock<Selector> = LazyLock::new(|| compile_selector("style"));
-/// Selector matching every descendant of `<body>`, used by hidden-element
-/// detection in Task 7.
-static SEL_BODY_ALL: LazyLock<Selector> = LazyLock::new(|| compile_selector("body *"));
-
-/// Shared ammonia builder. Configuration lands in Task 10.
-static AMMONIA_BUILDER: LazyLock<Builder<'static>> = LazyLock::new(build_ammonia_builder);
-
-/// Build the ammonia `Builder` used for Sprint 4b html sanitization.
-///
-/// Restricts URL schemes to `{http, https, mailto, tel}` and locks
-/// `<img>` attributes to `{alt, width, height}`, dropping `src`,
-/// `srcset`, and other remote-fetching surfaces. Ammonia's defaults
-/// already strip `<script>`, `<style>`, and event handler attributes.
-///
-/// Uses `rm_tag_attributes` + `add_tag_attributes` rather than
-/// `tag_attributes` so the defaults for unrelated tags (notably
-/// `<a href>`, which Task 11 needs to enumerate post-sanitize)
-/// remain in place.
-fn build_ammonia_builder() -> Builder<'static> {
-    let mut builder = Builder::default();
-    let schemes: HashSet<&'static str> = ["http", "https", "mailto", "tel"].into_iter().collect();
-    builder.url_schemes(schemes);
-    builder.rm_tag_attributes("img", &["src", "srcset"]);
-    builder.add_tag_attributes("img", &["alt", "width", "height"]);
-    // Pin tag removals against ammonia default drift. details/summary
-    // are explicitly removed: collapsed content is invisible to humans
-    // but visible to LLMs reading HTML tokens.
-    builder.rm_tags(&[
-        "script", "style", "iframe", "object", "embed", "meta", "base", "link", "form", "input",
-        "button", "textarea", "svg", "math", "frame", "frameset", "noframes", "applet", "details",
-        "summary",
-    ]);
-    builder.strip_comments(true);
-    builder
-}
-
-/// Count elements in `document` matching `selector`.
-fn count_matching(document: &Html, selector: &Selector) -> usize {
-    document.select(selector).count()
-}
-
-/// Count `<img>` elements in `document` whose `src` attribute is
-/// present and non-empty after trimming.
-fn count_img_with_src(document: &Html) -> usize {
-    document
-        .select(&SEL_IMG)
-        .filter(|el| el.value().attr("src").is_some_and(|s| !s.trim().is_empty()))
-        .count()
-}
-
-/// Parse a single `style="..."` attribute value into lowercased
-/// `(property, value)` pairs.
-///
-/// Very permissive: declarations are split on `;`, then each
-/// declaration is split on its first `:`. Empty properties or values
-/// are dropped. The intent is "does this style contain X", not full
-/// CSS conformance.
-fn parse_inline_style(style: &str) -> Vec<(String, String)> {
-    let mut pairs = Vec::new();
-    for decl in style.split(';') {
-        let Some((prop, val)) = decl.split_once(':') else {
-            continue;
-        };
-        let prop = prop.trim().to_ascii_lowercase();
-        let val = val.trim().to_ascii_lowercase();
-        if prop.is_empty() || val.is_empty() {
-            continue;
-        }
-        pairs.push((prop, val));
-    }
-    pairs
-}
-
-/// Parse a CSS length like `-9999px` into a pixel count.
-///
-/// Returns `None` for non-pixel units (em, rem, %, etc.) — they are
-/// treated as non-offscreen by design (inline-style-only scope).
-fn parse_px(val: &str) -> Option<f64> {
-    let stripped = val.strip_suffix("px").unwrap_or(val);
-    stripped.trim().parse::<f64>().ok()
-}
-
-/// Return `true` when an `opacity` value parses to (approximately) zero.
-fn opacity_is_zero(val: &str) -> bool {
-    let stripped = val.trim_end_matches('%').trim();
-    stripped
-        .parse::<f64>()
-        .ok()
-        .is_some_and(|n| n <= f64::EPSILON)
-}
-
-/// Return `true` when a `font-size` value parses to (approximately) zero.
-fn font_size_is_zero(val: &str) -> bool {
-    let digits: String = val
-        .chars()
-        .take_while(|c| c.is_ascii_digit() || *c == '.')
-        .collect();
-    digits
-        .parse::<f64>()
-        .ok()
-        .is_some_and(|n| n <= f64::EPSILON)
-}
-
-/// Parse a `transform: translate*(-Npx)` value and return the most
-/// negative pixel offset found, or `None` if no translate pattern
-/// matches.
-fn parse_translate_px(val: &str) -> Option<f64> {
-    let mut min: Option<f64> = None;
-    for part in val.split(['(', ',', ')']) {
-        let trimmed = part.trim();
-        if let Some(px_val) = parse_px(trimmed) {
-            match min {
-                Some(current) if px_val < current => min = Some(px_val),
-                None => min = Some(px_val),
-                Some(_) => {}
-            }
-        }
-    }
-    min
-}
-
-/// Accumulator for off-screen / color-match detection across an inline
-/// style declaration list.
-#[derive(Default)]
-struct StyleHints {
-    position: Option<String>,
-    left_px: Option<f64>,
-    top_px: Option<f64>,
-    transform_offset_px: Option<f64>,
-    color: Option<String>,
-    bg_color: Option<String>,
-}
-
-impl StyleHints {
-    fn record(&mut self, prop: &str, val: &str) {
-        match prop {
-            "position" => self.position = Some(val.to_string()),
-            "left" => self.left_px = parse_px(val),
-            "top" => self.top_px = parse_px(val),
-            "transform" => self.transform_offset_px = parse_translate_px(val),
-            "color" => self.color = Some(val.to_string()),
-            "background-color" => self.bg_color = Some(val.to_string()),
-            _ => {}
-        }
-    }
-
-    fn is_offscreen(&self) -> bool {
-        let positioned = self
-            .position
-            .as_deref()
-            .is_some_and(|p| p == "absolute" || p == "fixed");
-        if !positioned {
-            return false;
-        }
-        let off_left = self.left_px.is_some_and(|v| v < -100.0);
-        let off_top = self.top_px.is_some_and(|v| v < -100.0);
-        let off_transform = self.transform_offset_px.is_some_and(|v| v < -100.0);
-        off_left || off_top || off_transform
-    }
-
-    fn is_color_match(&self) -> bool {
-        match (self.color.as_ref(), self.bg_color.as_ref()) {
-            (Some(c), Some(bg)) => c == bg,
-            (Some(_) | None, None) | (None, Some(_)) => false,
-        }
-    }
-}
-
-/// Check a single declaration for an immediate hidden-method match.
-///
-/// Returns `Some` only for self-contained patterns (display, visibility,
-/// opacity, font-size). Multi-property patterns (off-screen, color
-/// match) accumulate via [`StyleHints`] and are resolved by the caller.
-fn classify_single_declaration(prop: &str, val: &str) -> Option<HiddenMethod> {
-    match prop {
-        "display" if val == "none" => Some(HiddenMethod::DisplayNone),
-        "visibility" if val == "hidden" => Some(HiddenMethod::VisibilityHidden),
-        "opacity" if opacity_is_zero(val) => Some(HiddenMethod::OpacityZero),
-        "font-size" if font_size_is_zero(val) => Some(HiddenMethod::ZeroFont),
-        _ => None,
-    }
-}
-
-/// Classify an inline `style` string into a [`HiddenMethod`], if any.
-fn classify_inline_style(style: &str) -> Option<HiddenMethod> {
-    let pairs = parse_inline_style(style);
-    let mut hints = StyleHints::default();
-    for (prop, val) in &pairs {
-        if let Some(method) = classify_single_declaration(prop, val) {
-            return Some(method);
-        }
-        hints.record(prop, val);
-    }
-    if hints.is_offscreen() {
-        return Some(HiddenMethod::OffScreen);
-    }
-    if hints.is_color_match() {
-        return Some(HiddenMethod::ColorMatch);
-    }
-    None
-}
-
-/// Extract the registrable domain from a URL-looking string.
-///
-/// Returns `None` for empty input, relative URLs, `mailto:`/`tel:`/
-/// `javascript:`/`data:` schemes, single-label hosts, and any input the
-/// PSL parser cannot resolve to a registrable domain.
-fn extract_registrable_domain(url_or_host: &str) -> Option<String> {
-    let trimmed = url_or_host.trim();
-    if trimmed.is_empty() {
-        return None;
-    }
-    let lowered = trimmed.to_ascii_lowercase();
-    if lowered.starts_with("mailto:")
-        || lowered.starts_with("tel:")
-        || lowered.starts_with("javascript:")
-        || lowered.starts_with("data:")
-    {
-        return None;
-    }
-    let after_scheme = lowered
-        .split_once("://")
-        .map_or(lowered.as_str(), |(_, rest)| rest);
-    let host = after_scheme
-        .split(['/', '?', '#'])
-        .next()
-        .unwrap_or("")
-        .trim_start_matches("www.");
-    let host = host.split(':').next().unwrap_or(host);
-    if host.is_empty() || !host.contains('.') {
-        return None;
-    }
-    let ascii = idna::domain_to_ascii(host).ok()?;
-    let domain = addr::parse_domain_name(ascii.as_str()).ok()?;
-    Some(domain.root()?.to_string())
-}
-
-/// A single href-mismatch hit recorded by [`detect_mismatches`].
-#[derive(Debug, Clone)]
-struct MismatchHit {
-    text_domain: String,
-    href_domain: String,
-}
-
-/// Walk every `<a href>` and report cases where a URL-looking token in
-/// the anchor text resolves to a different registrable domain than the
-/// `href` attribute.
-///
-/// Returns `(hits, overflow)`. `hits` contains at most
-/// [`MAX_MISMATCH_HITS`] entries; `overflow` counts additional mismatches
-/// past the cap so the caller can emit a summary warning.
-fn detect_mismatches(document: &Html) -> (Vec<MismatchHit>, usize, Vec<(String, String)>) {
-    let mut hits = Vec::new();
-    let mut overflow: usize = 0;
-    let mut unparsable_hrefs: Vec<(String, String)> = Vec::new();
-    let mut finder = linkify::LinkFinder::new();
-    finder.url_must_have_scheme(false);
-    for anchor in document.select(&SEL_ANCHOR) {
-        let Some(href) = anchor.value().attr("href") else {
-            continue;
-        };
-        let Some(href_domain) = extract_registrable_domain(href) else {
-            let mut text: String = anchor.text().collect::<Vec<&str>>().join(" ");
-            if text.len() > MAX_ANCHOR_TEXT_SCAN {
-                text.truncate(MAX_ANCHOR_TEXT_SCAN);
-            }
-            let has_url_text = finder
-                .links(&text)
-                .any(|l| l.kind() == &linkify::LinkKind::Url);
-            if has_url_text {
-                unparsable_hrefs.push((href.to_string(), text.trim().to_string()));
-            }
-            continue;
-        };
-        let mut text: String = anchor.text().collect::<Vec<&str>>().join(" ");
-        if text.len() > MAX_ANCHOR_TEXT_SCAN {
-            text.truncate(MAX_ANCHOR_TEXT_SCAN);
-        }
-        let mut link_iter = finder
-            .links(&text)
-            .filter(|l| l.kind() == &linkify::LinkKind::Url);
-        let Some(link) = link_iter.next() else {
-            continue;
-        };
-        let Some(text_domain) = extract_registrable_domain(link.as_str()) else {
-            continue;
-        };
-        if text_domain.eq_ignore_ascii_case(&href_domain) {
-            continue;
-        }
-        if hits.len() < MAX_MISMATCH_HITS {
-            hits.push(MismatchHit {
-                text_domain,
-                href_domain,
-            });
-        } else {
-            overflow += 1;
-        }
-    }
-    (hits, overflow, unparsable_hrefs)
-}
-
-/// Walk the document and collect hidden-element hits plus their
-/// tree-order indices (so text extraction can skip them later).
-///
-/// Returns `(hits, overflow)`. `hits` contains at most
-/// [`MAX_HIDDEN_HITS`] entries; `overflow` is the count of additional
-/// hidden elements detected past the cap so the caller can emit a
-/// summary warning.
-fn detect_hidden(document: &Html) -> (Vec<(ElementIndex, HiddenMethod)>, usize) {
-    let mut hits = Vec::new();
-    let mut overflow: usize = 0;
-    for (idx, element) in document.select(&SEL_BODY_ALL).enumerate() {
-        let Some(style) = element.value().attr("style") else {
-            continue;
-        };
-        let Some(method) = classify_inline_style(style) else {
-            continue;
-        };
-        if hits.len() < MAX_HIDDEN_HITS {
-            hits.push((idx, method));
-        } else {
-            overflow += 1;
-        }
-    }
-    (hits, overflow)
-}
-
-/// Extract plain text from the document, skipping hidden elements
-/// and non-content tags (`<script>`, `<style>`, `<noscript>`,
-/// `<template>`, `<head>`, `<title>`).
-///
-/// `hidden_indices` is the set of element indices produced by
-/// [`detect_hidden`]; it is consulted during a pre-order recursion
-/// over `<body>`'s descendants. The recursion increments a shared
-/// counter once per element child encountered, matching the
-/// enumeration order of `select(&SEL_BODY_ALL)` so the two index
-/// spaces stay aligned.
-///
-/// The collected buffer is whitespace-normalized via
-/// [`normalize_whitespace`] and then routed through
-/// [`crate::unicode::sanitize`] to share the unicode pipeline used by
-/// the rest of `rimap-content`. Any warnings produced by the
-/// sanitizer are returned alongside the text so the caller can merge
-/// them into the [`HtmlResult`] warnings list.
-fn extract_text(
-    document: &Html,
-    hidden_indices: &HashSet<ElementIndex>,
-) -> (String, Vec<SecurityWarning>) {
-    let mut buf = String::new();
-    let body_selector = compile_selector("body");
-    if let Some(body_el) = document.select(&body_selector).next() {
-        let mut counter: usize = 0;
-        walk_children(body_el, hidden_indices, &mut buf, &mut counter);
-    }
-    let normalized = normalize_whitespace(&buf);
-    crate::unicode::sanitize(
-        normalized.as_bytes(),
-        Some("utf-8"),
-        MAX_HTML_BYTES,
-        "body:html",
-    )
-}
-
-/// Walk `children` of an element, recursing into each element child via
-/// [`walk_element`] and appending plain text. CDATA comments suppress
-/// the next adjacent text node so CDATA marker bookkeeping does not
-/// leak into the output. Shared between the body-root loop and the
-/// recursive element walker to keep counting semantics identical.
-fn walk_children(
-    parent: scraper::ElementRef<'_>,
-    hidden_indices: &HashSet<ElementIndex>,
-    out: &mut String,
-    counter: &mut usize,
-) {
-    let mut after_cdata = false;
-    for child in parent.children() {
-        if let Some(child_el) = scraper::ElementRef::wrap(child) {
-            after_cdata = false;
-            walk_element(child_el, hidden_indices, out, counter);
-        } else if child
-            .value()
-            .as_comment()
-            .is_some_and(|c| c.starts_with("[CDATA["))
-        {
-            after_cdata = true;
-        } else if let Some(text) = child.value().as_text()
-            && !after_cdata
-        {
-            push_text(out, text);
-        }
-    }
-}
-
-/// Recursive helper for [`extract_text`].
-///
-/// Visits `el` (already counted by the caller against
-/// `hidden_indices`), short-circuiting on non-content tags, then
-/// walks its children: element children recurse via this function,
-/// text children are appended to `out`. Each element child increments
-/// `counter` exactly once before its hidden-skip check, mirroring the
-/// pre-order enumeration in [`detect_hidden`].
-fn collect_visible_text(
-    el: scraper::ElementRef<'_>,
-    hidden_indices: &HashSet<ElementIndex>,
-    out: &mut String,
-    counter: &mut usize,
-) {
-    let tag = el.value().name();
-    if NON_CONTENT_TAGS.contains(&tag) {
-        return;
-    }
-    walk_children(el, hidden_indices, out, counter);
-}
-
-/// HTML tags whose contents are not considered user-visible body text
-/// and are therefore skipped during hidden-text extraction. Named slice
-/// avoids both `matches!` (banned by project style) and a wildcard
-/// match arm against a non-enum `&str` input.
-const NON_CONTENT_TAGS: &[&str] = &["script", "style", "noscript", "template", "head", "title"];
-
-/// Increment the counter for `child_el`, skip if hidden, otherwise
-/// recurse via [`collect_visible_text`]. Extracted so the body-root
-/// loop and the recursive walker share identical counting semantics.
-fn walk_element(
-    child_el: scraper::ElementRef<'_>,
-    hidden_indices: &HashSet<ElementIndex>,
-    out: &mut String,
-    counter: &mut usize,
-) {
-    let my_idx = *counter;
-    *counter += 1;
-    if hidden_indices.contains(&my_idx) {
-        return;
-    }
-    collect_visible_text(child_el, hidden_indices, out, counter);
-}
-
-/// Append a text node's contents to `out`, inserting a separating
-/// space when the buffer is non-empty and does not already end in
-/// whitespace. Internal whitespace is left intact for
-/// [`normalize_whitespace`] to collapse.
-///
-/// Text nodes containing `]]>` are skipped as a secondary defense
-/// against CDATA leaks. html5ever treats `<![CDATA[` in non-SVG/
-/// `MathML` context as a bogus comment; content between inner tags
-/// and `]]>` leaks as text nodes. The primary defense is the
-/// `after_cdata` flag in [`extract_text`] and
-/// [`collect_visible_text`], which suppresses text siblings that
-/// immediately follow a CDATA bogus-comment node (covering the
-/// unclosed-CDATA case where `]]>` is absent).
-fn push_text(out: &mut String, text: &str) {
-    if text.contains("]]>") {
-        return;
-    }
-    if !out.is_empty() && !out.ends_with(char::is_whitespace) {
-        out.push(' ');
-    }
-    out.push_str(text);
-}
-
-/// Collapse runs of ASCII/Unicode whitespace in `s` to single spaces
-/// and trim leading/trailing whitespace.
-fn normalize_whitespace(s: &str) -> String {
-    let mut out = String::with_capacity(s.len());
-    let mut prev_space = false;
-    for c in s.chars() {
-        if c.is_whitespace() {
-            if !prev_space && !out.is_empty() {
-                out.push(' ');
-            }
-            prev_space = true;
-        } else {
-            out.push(c);
-            prev_space = false;
-        }
-    }
-    let trimmed = out.trim_end();
-    trimmed.to_string()
-}
-
 /// Process a raw HTML body into sanitized text + html + warnings.
 ///
 /// # Arguments
@@ -608,17 +117,6 @@ fn normalize_whitespace(s: &str) -> String {
 /// Returns [`ContentError::LimitExceeded`] if `raw` exceeds
 /// [`MAX_HTML_BYTES`].
 pub(crate) fn process(raw: &[u8], charset: Option<&str>) -> Result<HtmlResult, ContentError> {
-    // Warm the LazyLocks so the const selectors and ammonia builder are
-    // exercised on first call. This both validates them at runtime and
-    // keeps them out of dead-code analysis until later tasks consume them.
-    let _ = (
-        &*SEL_ANCHOR,
-        &*SEL_IMG,
-        &*SEL_SCRIPT,
-        &*SEL_STYLE,
-        &*SEL_BODY_ALL,
-        &*AMMONIA_BUILDER,
-    );
     if raw.len() > MAX_HTML_BYTES {
         return Err(ContentError::LimitExceeded {
             kind: HTML_BODY_LIMIT_KIND,
@@ -681,66 +179,15 @@ pub(crate) fn process(raw: &[u8], charset: Option<&str>) -> Result<HtmlResult, C
     })
 }
 
-/// Stage 8: re-parse the ammonia-sanitized `body_html` and collect every
-/// surviving `<a href>` value in document order. Consumed by
-/// `lookalike::audit` (Sprint 4b Task 15).
-///
-/// The re-parse is intentional: ammonia's output is the canonical
-/// "what the recipient might click", and we want anchor hrefs that
-/// reflect post-sanitization reality (so e.g. `javascript:` URLs
-/// stripped in Task 10's allowlist do not appear here).
-fn collect_anchor_hrefs(sanitized_html: &str) -> Vec<String> {
-    let doc = Html::parse_document(sanitized_html);
-    doc.select(&SEL_ANCHOR)
-        .filter_map(|a| a.value().attr("href").map(str::to_string))
-        .collect()
-}
-
-/// Stage 7: count pre-sanitize remote-content elements on the
-/// scraper-parsed `document`, run `ammonia::clean` on `decoded`, and
-/// emit `HtmlScriptStripped` / `HtmlStyleStripped` /
-/// `HtmlRemoteImageStripped` warnings when the pre-sanitize count is
-/// non-zero. Returns the sanitized HTML string.
-///
-/// The counting deliberately runs against the same `Html` value used
-/// by all earlier detection stages (html5ever 0.39 via scraper), not
-/// against ammonia's internal parse (html5ever 0.35). This means a
-/// crafted divergence between the two tokenizers is observable as a
-/// warning-count vs. `body_html` mismatch — see the Task 17 corpus.
-fn sanitize_body(document: &Html, decoded: &str, warnings: &mut Vec<SecurityWarning>) -> String {
-    let script_count = count_matching(document, &SEL_SCRIPT);
-    let style_count = count_matching(document, &SEL_STYLE);
-    let remote_img_count = count_img_with_src(document);
-    let body_html = AMMONIA_BUILDER.clean(decoded).to_string();
-    if script_count > 0 {
-        warnings.push(SecurityWarning::at(
-            crate::output::WarningCode::HtmlScriptStripped,
-            format!("count={script_count}"),
-            "body:html",
-        ));
-    }
-    if style_count > 0 {
-        warnings.push(SecurityWarning::at(
-            crate::output::WarningCode::HtmlStyleStripped,
-            format!("count={style_count}"),
-            "body:html",
-        ));
-    }
-    if remote_img_count > 0 {
-        warnings.push(SecurityWarning::at(
-            crate::output::WarningCode::HtmlRemoteImageStripped,
-            format!("count={remote_img_count}"),
-            "body:html",
-        ));
-    }
-    body_html
-}
-
 #[cfg(test)]
 #[expect(clippy::expect_used, reason = "tests may expect on constructed values")]
 #[expect(clippy::panic, reason = "test failure paths")]
 mod tests {
     use super::*;
+    use crate::html::hidden::{SEL_BODY_ALL, compile_selector};
+    use crate::html::mismatch::{SEL_ANCHOR, SEL_IMG, extract_registrable_domain};
+    use crate::html::sanitize::{AMMONIA_BUILDER, SEL_SCRIPT, SEL_STYLE};
+    use crate::html::style_parse::{classify_inline_style, classify_single_declaration};
 
     #[test]
     fn process_oversize_input_returns_limit_exceeded() {
@@ -794,9 +241,6 @@ mod tests {
 
     #[test]
     fn constants_are_referenced() {
-        // The size, anchor-scan, and hit caps will be consumed by
-        // Tasks 7/8/10. Reference them here so test builds don't trip
-        // dead-code while the production call sites are still stubs.
         let _ = (
             MAX_HTML_BYTES,
             MAX_ANCHOR_TEXT_SCAN,
@@ -885,13 +329,6 @@ mod tests {
     }
 
     /// Negative cases for [`classify_single_declaration`] guards.
-    ///
-    /// These assertions kill mutants that weaken the `val == "none"`,
-    /// `val == "hidden"`, and `font_size_is_zero(val)` match guards to
-    /// `true`: with the guard removed, any value for `display`,
-    /// `visibility`, or `font-size` would classify as hidden, so a
-    /// non-matching value that still uses the property must return
-    /// `None`.
     #[test]
     fn classify_single_declaration_visible_values_return_none() {
         assert_eq!(classify_single_declaration("display", "block"), None);
@@ -904,9 +341,7 @@ mod tests {
         assert_eq!(classify_single_declaration("color", "red"), None);
     }
 
-    /// Positive cases for [`classify_single_declaration`] that pin the
-    /// variant produced by each matching guard. These kill mutants that
-    /// swap `DisplayNone`/`VisibilityHidden`/`OpacityZero`/`ZeroFont`.
+    /// Positive cases for [`classify_single_declaration`].
     #[test]
     fn classify_single_declaration_variant_per_property() {
         assert_eq!(
@@ -991,10 +426,6 @@ mod tests {
             .find(|w| matches!(w.code, crate::output::WarningCode::HtmlLinkTextHrefMismatch))
             .expect("expected mismatch warning");
         let detail = mismatch.detail.as_deref().expect("detail present");
-        // Detail records the registrable (PSL root), so `bank.example.com`
-        // collapses to `example.com` here. The plan-text spec asserted the
-        // raw subdomain, but that contradicts the matching-subdomain test
-        // and the documented behavior of `extract_registrable_domain`.
         assert!(detail.contains("text_domain=example.com"), "got {detail}");
         assert!(
             detail.contains("href_domain=attacker.example"),
@@ -1034,10 +465,6 @@ mod tests {
         );
     }
 
-    /// Unit coverage for [`extract_registrable_domain`]'s scheme-skip
-    /// short-circuits. Each non-web scheme must independently return
-    /// `None`; this kills mutants that convert any of the chained `||`
-    /// checks into `&&`.
     #[test]
     fn extract_registrable_domain_skips_non_web_schemes() {
         assert_eq!(extract_registrable_domain("mailto:foo@example.com"), None);
@@ -1050,9 +477,6 @@ mod tests {
         assert_eq!(extract_registrable_domain("singlelabel"), None);
     }
 
-    /// Pins the positive behaviour of [`extract_registrable_domain`] so
-    /// that mutants flipping boundary comparisons or dropping scheme
-    /// parsing are caught.
     #[test]
     fn extract_registrable_domain_returns_psl_root() {
         assert_eq!(
@@ -1152,11 +576,6 @@ mod tests {
 
     #[test]
     fn extract_text_index_alignment_skips_only_hidden_elements() {
-        // Three sibling spans, the middle one hidden via display:none.
-        // Visible siblings on either side must survive; the hidden one
-        // and its text must not. This pins the index alignment between
-        // detect_hidden's SEL_BODY_ALL enumeration and extract_text's
-        // pre-order recursion.
         let input = br#"<html><body>
             <span>alpha</span>
             <span style="display:none">SECRET</span>
@@ -1182,10 +601,6 @@ mod tests {
 
     #[test]
     fn extract_text_index_alignment_handles_nested_hidden() {
-        // Hidden element with a visible-text descendant: the entire
-        // hidden subtree must be omitted. A later visible sibling at a
-        // larger index confirms the counter advanced past the skipped
-        // descendants in lock-step with detect_hidden.
         let input = br#"<html><body>
             <p>before</p>
             <div style="display:none"><span>nested hidden</span><em>still hidden</em></div>
@@ -1267,8 +682,6 @@ mod tests {
             <a href="javascript:alert(1)">bad</a>
         </body></html>"#;
         let result = process(input, None).expect("ok");
-        // javascript: URL was stripped by ammonia in Task 10, so only 3
-        // survive in the sanitized HTML.
         assert_eq!(result.anchor_hrefs.len(), 3);
         assert!(
             result
@@ -1331,8 +744,6 @@ mod tests {
 
     #[test]
     fn lazylocks_initialize_without_panic() {
-        // Touch each LazyLock so the compile_selector expectation is
-        // exercised even if the process() warming pattern changes later.
         let _ = &*SEL_ANCHOR;
         let _ = &*SEL_IMG;
         let _ = &*SEL_SCRIPT;
